Move getRepeatableJobsByID to workers/helpers.js
